+++ src/resources/views/declaration.blade.php
@@ -204,22 +204,6 @@
                                             @endif
                                         </td>
                                     </tr>
-{{--                                    <tr>--}}
-{{--                                        <td>--}}
-{{--                                            {{ __('app.City from and') }}:&nbsp;--}}
-{{--                                            <strong class="text-uppercase">--}}
-{{--                                                {{ $declaration['travelling_from_city'] }}--}}
-{{--                                            </strong>--}}
-{{--                                        </td>--}}
-{{--                                    </tr>--}}
-{{--                                    <tr>--}}
-{{--                                        <td>--}}
-{{--                                            {{ __('app.Date travelling') }}:--}}
-{{--                                            &nbsp;{{ __('app.Year') }} <strong>{{ $declaration['travelling_date_year'] }}</strong>,--}}
-{{--                                            &nbsp;{{ __('app.Month') }} <strong>{{ $declaration['travelling_date_month'] }}</strong>,--}}
-{{--                                            &nbsp;{{ __('app.Day') }} <strong>{{ $declaration['travelling_date_day'] }}</strong>--}}
-{{--                                        </td>--}}
-{{--                                    </tr>--}}
                                 </table>
                             </div>
                         </div>
@@ -241,16 +225,6 @@
                                         ____________________________________________________________________,
                                     @endif
                                 </p>
-{{--                                <p class="no-margin-bottom">--}}
-{{--                                    <span class="bullet-padding-right">&#8226;</span>--}}
-{{--                                    {!! __('app.I will travel by') !!}:&nbsp;--}}
-{{--                                    @if (strlen($declaration['vehicle_registration_no']) > 0)--}}
-{{--                                        {{ __('app.' . $declaration['vehicle_type']) }}--}}
-{{--                                        <strong>{{ $declaration['vehicle_registration_no'] }}</strong>--}}
-{{--                                    @else--}}
-{{--                                        _______________________________--}}
-{{--                                    @endif--}}
-{{--                                </p>--}}
                                 <p class="no-margin-bottom">
                                     <span class="bullet-padding-right">&#8226;</span>
                                     {!! __('app.I agree to the use of my personal data') !!}:
@@ -262,7 +236,6 @@
                                 <p class="no-margin-bottom">
                                     {!! __('app.During my stay in Romania I can be reached at') !!}:<br />
                                     {{ __('app.Table Phone') }}: <strong>{{ $declaration['phone'] }}</strong><br />
-{{--                                    {{ __('app.Table E-mail') }}: <strong>{{ $declaration['email'] }}</strong>--}}
                                 </p>
                             </div>
                         </div>
@@ -292,11 +265,6 @@
                                     <tr>
                                         <td>
                                             {{ $declaration['current_date'] }}
-{{--                                            @if (strlen($declaration['border']) > 0)--}}
-{{--                                                {{ $declaration['border'] }}--}}
-{{--                                            @else--}}
-{{--                                                ________________________________--}}
-{{--                                            @endif--}}
                                         </td>
                                     </tr>
                                 </table>